requirements.txt
faiss-gpu==1.7.2
imageio==2.31.1
imutils==0.5.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.4
onnxruntime==1.16.3
opencv-python==4.7.0.72
Pillow==9.5.0
pytorch-metric-learning==2.4.1
record-keeper==0.9.32
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.10.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==1.13.0
torchvision==0.14.0
umap-learn==0.5.5
uri-template==1.3.0